Split SignupSubscribeToCurated component
